+++ CHANGELOG.md
@@ -5,8 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] - 2024-03-25
+
+### Improvement
+
+- Set default values for the database object.
+- Expose the computed database configuration.
+- Limit `min_capacity` and `auto_pause_delay_in_minutes` to Serverless SKUs.
+- Improve variables and outputs descriptions.
+
 ## [1.0.0] - 2024-08-01
 
 ### Added
 
-- Initial Release to open source
+- Initial Release to Open Source